#11793 Delete User Dependency before delete the User https://github.com/metasfresh/metasfresh/issues/11793 - prevent delete logged-in user

+++ backend/de.metas.adempiere.adempiere/base/src/main/java/de/metas/user/api/impl/UserBL.java
@@ -12,6 +12,7 @@ import de.metas.email.mailboxes.UserEMailConfig;
 import de.metas.email.templates.MailTemplateId;
 import de.metas.email.templates.MailTextBuilder;
 import de.metas.i18n.AdMessageKey;
+import de.metas.i18n.IMsgBL;
 import de.metas.i18n.ITranslatableString;
 import de.metas.i18n.Language;
 import de.metas.i18n.TranslatableStrings;
@@ -53,6 +54,8 @@ import java.util.UUID;
 
 public class UserBL implements IUserBL
 {
+	private static final AdMessageKey MSG_UserDelete = AdMessageKey.of("UserDeleteMsg");
+
 	private static final Logger logger = LogManager.getLogger(UserBL.class);
 	private final IUserDAO userDAO = Services.get(IUserDAO.class);
 	private final IClientDAO clientDAO = Services.get(IClientDAO.class);
@@ -61,7 +64,7 @@ public class UserBL implements IUserBL
 	private final ISysConfigBL sysConfigBL = Services.get(ISysConfigBL.class);
 	private final IValuePreferenceDAO valuePreferenceDAO = Services.get(IValuePreferenceDAO.class);
 	private final IRoleDAO roleDAO = Services.get(IRoleDAO.class);
-
+	private final IMsgBL msgBL = Services.get(IMsgBL.class);
 	/**
 	 * @see org.compiere.model.X_AD_MailConfig#CUSTOMTYPE_OrgCompiereUtilLogin
 	 */
@@ -478,22 +481,35 @@ public class UserBL implements IUserBL
 	@Override
 	public void deleteUserDependency(@NonNull final I_AD_User userRecord)
 	{
-		UserId userId = UserId.ofRepoId(userRecord.getAD_User_ID());
+		if (Env.getAD_User_ID() != userRecord.getAD_User_ID())
+		{
+			UserId userId = UserId.ofRepoId(userRecord.getAD_User_ID());
 
-		valuePreferenceDAO.deleteUserPreferenceByUserId(userId);
-		getUserAuthTokenRepository().deleteUserAuthTokenByUserId(userId);
+			valuePreferenceDAO.deleteUserPreferenceByUserId(userId);
+			getUserAuthTokenRepository().deleteUserAuthTokenByUserId(userId);
 
-		userRolePermissionsDAO.deleteUserOrgAccessByUserId(userId);
+			userRolePermissionsDAO.deleteUserOrgAccessByUserId(userId);
 
-		userRolePermissionsDAO.deleteUserOrgAssignmentByUserId(userId);
+			userRolePermissionsDAO.deleteUserOrgAssignmentByUserId(userId);
 
-		roleDAO.deleteUserRolesByUserId(userId);
+			roleDAO.deleteUserRolesByUserId(userId);
 
-		getUserSubstituteRepository().deleteUserSubstituteByUserId(userId);
+			getUserSubstituteRepository().deleteUserSubstituteByUserId(userId);
 
-		getUserMailRepository().deleteUserMailByUserId(userId);
+			getUserMailRepository().deleteUserMailByUserId(userId);
 
-		getUserQueryRepository().deleteUserQueryByUserId(userId);
+			getUserQueryRepository().deleteUserQueryByUserId(userId);
+		}
+		else
+		{
+			final ITranslatableString errorMsg = msgBL.getTranslatableMsgText(MSG_UserDelete);
+			throw new AdempiereException(TranslatableStrings.builder()
+												 .append(errorMsg)
+												 .build())
+					.appendParametersToMessage()
+					.setParameter("AD_User_ID", userRecord.getAD_User_ID())
+					.setParameter("Name", userRecord.getName());
+		}
 	}
 
 }
